SMV source: make changes to remove INTEL compiler warnings
Former-commit-id: 7857dbffe30bc9c5bb70c1d5bebab9a301b641e4 [formerly eb3122f997db11b2ec2e8d578de176b71398e871]
Former-commit-id: 584591bdedc883bac16c5829eac7ece044b416af
Former-commit-id: 58d93cd7ebfc3507abe2cf196fdd9184cf6f1503

+++ sv_5/source/smokeview/IOiso.c
@@ -1516,7 +1516,7 @@ void uncompress_isodataframe(isosurface *asurface, int n){
     uncompress(full_data,&countout,compressed_data,countin);
     if(countout!= asurface->nfull_bufferframe){
       printf("problems with decompressing iso data\n");
-      printf(" %i %i\n",asurface->nfull_bufferframe,countout);
+      printf(" %i %i\n",asurface->nfull_bufferframe,(int)countout);
     }
   }
